requirements.txt
# Elasticsearch/Kibana client
elasticsearch>=8.0.0

# Encryption libraries
cryptography>=41.0.0
pycryptodome>=3.19.0

# CLI and utilities
click>=8.1.0
python-dotenv>=1.0.0
requests>=2.31.0

# Data handling
python-dateutil>=2.8.0

# Pretty printing
rich>=13.0.0
